% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam.R
\docType{data}
\name{adam_adadas}
\alias{adam_adadas}
\title{adam_adadas}
\format{
a data frame with 12463 rows and 40 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/adam/README.md} for more details and full data specification.
}
\usage{
adam_adadas
}
\description{
ADaM adadas data frame
}
\keyword{datasets}
