% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam.R
\docType{data}
\name{adam_adcibc}
\alias{adam_adcibc}
\title{adam_adcibc}
\format{
a data frame with 730 rows and 36 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/adam/README.md} for more details and full data specification.
}
\usage{
adam_adcibc
}
\description{
ADaM adcibc data frame
}
\keyword{datasets}
