% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam.R
\docType{data}
\name{adam_adlbcpv}
\alias{adam_adlbcpv}
\title{adam_adlbcpv}
\format{
a data frame with 37132 rows and 46 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/adam/README.md} for more details and full data specification.
}
\usage{
adam_adlbcpv
}
\description{
ADaM adlbcpv data frame
}
\keyword{datasets}
