% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam.R
\docType{data}
\name{adam_adlbhpv}
\alias{adam_adlbhpv}
\title{adam_adlbhpv}
\format{
a data frame with 24966 rows and 46 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/adam/README.md} for more details and full data specification.
}
\usage{
adam_adlbhpv
}
\description{
ADaM adlbhpv data frame
}
\keyword{datasets}
