% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam.R
\docType{data}
\name{adam_adnpix}
\alias{adam_adnpix}
\title{adam_adnpix}
\format{
a data frame with 31140 rows and 41 columns. See \url{https://github.com/SafetyGraphics/safetyData/blob/main/data-raw/adam/README.md} for more details and full data specification.
}
\usage{
adam_adnpix
}
\description{
ADaM adnpix data frame
}
\keyword{datasets}
