% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{ae_counts}
\alias{ae_counts}
\title{ae_counts
Count of subjects with an adverse event}
\usage{
ae_counts(.data, ..., N_counts = NULL, sort = FALSE)
}
\description{
ae_counts
Count of subjects with an adverse event
}
