% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_plot.R
\name{AEplot}
\alias{AEplot}
\title{Create AEplot}
\usage{
AEplot(data, paramVar, aeStartVar, aeEndVar, colorVar)
}
\arguments{
\item{data}{AE data frame}

\item{paramVar}{AE Preferred Term Column AETERM}

\item{aeStartVar}{AE Start day column, ASTDY}

\item{aeEndVar}{AE End day column, AENDY}

\item{colorVar}{AE Severity, AESEV}
}
\value{
an AE  plot created with ggplot
}
\description{
Create AEplot
}
