% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{cmh_p}
\alias{cmh_p}
\title{cmh_p
CMH test p-value with options for alternate hyptheses}
\usage{
cmh_p(.data, formula, alternate = c("rmeans", "cmeans", "general", "cor"))
}
\description{
cmh_p
CMH test p-value with options for alternate hyptheses
}
