% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datarx.pilotADaM-package.R
\docType{package}
\name{datarx.pilotADaM-package}
\alias{datarx.pilotADaM}
\alias{datarx.pilotADaM-package}
\title{datarx.pilotADaM: Inspired by the CDISC_pilot_replication project, convert it to a whole new R package, and re-implment all the tables}
\description{
Inspired by the CDISC_pilot_replication project, SafetyData project and PHUSE TestDataFactory, migrate the CDISC data into a R package and re-implement the tables defined int he cdiscpilot project. A shiny example is created to demo the tables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/palandatarxcom/datarx.pilotADaM}
  \item Report bugs at \url{https://github.com/palandatarxcom/datarx.pilotADaM/issues}
}

}
\keyword{internal}
