% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{desc_stats}
\alias{desc_stats}
\title{desc_stats}
\usage{
desc_stats(
  .data,
  var,
  group = TRT01PN,
  na.rm = TRUE,
  int_len = 3,
  size = 10,
  include = c("n", "Mean", "SD", "Median", "Min", "Max")
)
}
\arguments{
\item{include}{}
}
\description{
desc_stats
}
