% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{efficacy_models}
\alias{efficacy_models}
\title{efficacy_models}
\usage{
efficacy_models(data, var = NULL, wk = NULL, model_type = "ancova")
}
\arguments{
\item{model_type}{}
}
\description{
efficacy_models
}
