% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{example_custom_reader}
\alias{example_custom_reader}
\title{example_custom_reader}
\usage{
example_custom_reader(..., table_number = NULL)
}
\arguments{
\item{table_number}{}
}
\description{
example_custom_reader
}
