% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{fisher_test_ae}
\alias{fisher_test_ae}
\title{fisher_test_ae
Fisher test built for row-wise derivations suited for our AE tables}
\usage{
fisher_test_ae(.data)
}
\description{
fisher_test_ae
Fisher test built for row-wise derivations suited for our AE tables
}
