% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTables.R
\name{gen_ht_t14_101}
\alias{gen_ht_t14_101}
\title{gen_ht_t14_101
Generate Table 14-1.01 Summary of Population}
\usage{
gen_ht_t14_101(adsl)
}
\arguments{
\item{adsl}{ADaM ADSL data set}
}
\value{
a huxtable table
}
\description{
gen_ht_t14_101
Generate Table 14-1.01 Summary of Population
}
