% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTables.R
\name{gen_ht_t14_102}
\alias{gen_ht_t14_102}
\title{gen_ht_t14_102
Generate Table 14-1.02 Summary of End of Study Data}
\usage{
gen_ht_t14_102(adsl)
}
\arguments{
\item{adsl}{ADaM ADSL data set}
}
\value{
a huxtable table
}
\description{
gen_ht_t14_102
Generate Table 14-1.02 Summary of End of Study Data
}
\examples{
gen_ht_t14_102(adsl = adsl)

}
