% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTables.R
\name{gen_ht_t14_103}
\alias{gen_ht_t14_103}
\title{gen_ht_t14_103
Generate Table 14-1.03 Summary of Number of Subjects By Site}
\usage{
gen_ht_t14_103(adsl)
}
\arguments{
\item{adsl}{}
}
\value{
a huxtable table
}
\description{
gen_ht_t14_103
Generate Table 14-1.03 Summary of Number of Subjects By Site
}
