% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTables.R
\name{gen_ht_t14_201}
\alias{gen_ht_t14_201}
\title{gen_ht_t14_201
Generate Table 14-2.01 Summary of Demographic and Baseline Characteristics}
\usage{
gen_ht_t14_201(adsl)
}
\arguments{
\item{adsl}{ADaM ADSL dataset}
}
\value{
a huxtable table
}
\description{
gen_ht_t14_201
Generate Table 14-2.01 Summary of Demographic and Baseline Characteristics
}
\examples{
gen_ht_t14_201(adsl)

}
