% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTables.R
\name{gen_ht_t14_301}
\alias{gen_ht_t14_301}
\title{gen_ht_t14_301
Generate Table 14-3.01 Primary Endpoint Analysis: ADAS Cog (11) - Change from Baseline to Week 24 - LOCF}
\usage{
gen_ht_t14_301(adadas)
}
\arguments{
\item{adadas}{}
}
\value{
a huxtable
}
\description{
gen_ht_t14_301
Generate Table 14-3.01 Primary Endpoint Analysis: ADAS Cog (11) - Change from Baseline to Week 24 - LOCF
}
\examples{
gen_ht_t14_301(adadas = adadas)

}
