% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTables.R
\name{gen_ht_t14_302}
\alias{gen_ht_t14_302}
\title{gen_ht_t14_302
Generate Table 14-3.02 Primary Endpoint Analysis: CIBIC+ - Summary at Week 24 - LOCF}
\usage{
gen_ht_t14_302(adcibc)
}
\arguments{
\item{adcibc}{}
}
\value{
a huxtable
}
\description{
gen_ht_t14_302
Generate Table 14-3.02 Primary Endpoint Analysis: CIBIC+ - Summary at Week 24 - LOCF
}
