% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTables.R
\name{gen_ht_t14_304}
\alias{gen_ht_t14_304}
\title{gen_ht_t14_304
Generate Table 14-3.04 CIBIC+ - Summary at Week 8 - LOCF}
\usage{
gen_ht_t14_304(adcibc)
}
\arguments{
\item{adcibc}{}
}
\value{
a huxtable
}
\description{
gen_ht_t14_304
Generate Table 14-3.04 CIBIC+ - Summary at Week 8 - LOCF
}
