% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTables.R
\name{gen_ht_t14_305}
\alias{gen_ht_t14_305}
\title{gen_ht_t14_305
Generate Table 14-3.05 ADAS Cog (11) - Change from Baseline to Week 16 - LOCF}
\usage{
gen_ht_t14_305(adadas)
}
\arguments{
\item{adadas}{}
}
\value{
a huxtable
}
\description{
gen_ht_t14_305
Generate Table 14-3.05 ADAS Cog (11) - Change from Baseline to Week 16 - LOCF
}
