% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTables.R
\name{gen_ht_t14_306}
\alias{gen_ht_t14_306}
\title{gen_ht_t14_306
Generate Table 14-3.06 CIBIC+ - Summary at Week 16 - LOCF}
\usage{
gen_ht_t14_306(adcibc)
}
\arguments{
\item{adcibc}{}
}
\value{
a huxtable
}
\description{
gen_ht_t14_306
Generate Table 14-3.06 CIBIC+ - Summary at Week 16 - LOCF
}
