% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTables.R
\name{gen_ht_t14_307}
\alias{gen_ht_t14_307}
\title{gen_ht_t14_307
Generate Table 14-3.07 ADAS Cog (11) - Change from Baseline to Week 24 - Completers at Wk 24-Observed Cases-Windowed}
\usage{
gen_ht_t14_307(adadas)
}
\arguments{
\item{adadas}{}
}
\value{
a huxtable
}
\description{
gen_ht_t14_307
Generate Table 14-3.07 ADAS Cog (11) - Change from Baseline to Week 24 - Completers at Wk 24-Observed Cases-Windowed
}
