% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTables.R
\name{gen_ht_t14_308}
\alias{gen_ht_t14_308}
\title{gen_ht_t14_308
Generate Table 14-3.08 ADAS Cog (11) - Change from Baseline to Week 24 in Male Subjects - LOCF}
\usage{
gen_ht_t14_308(adadas)
}
\value{
a huxtable
}
\description{
#' @param adadas
}
