% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTables.R
\name{gen_ht_t14_309}
\alias{gen_ht_t14_309}
\title{gen_ht_t14_309
Generate Table 14-3.09 ADAS Cog (11) - Change from Baseline to Week 24 in Female Subjects - LOCF}
\usage{
gen_ht_t14_309(adadas)
}
\arguments{
\item{adadas}{}
}
\description{
gen_ht_t14_309
Generate Table 14-3.09 ADAS Cog (11) - Change from Baseline to Week 24 in Female Subjects - LOCF
}
