% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTables.R
\name{gen_ht_t14_310}
\alias{gen_ht_t14_310}
\title{gen_ht_t14_310
Generate Table 14-3.10 ADAS Cog (11) - Mean and Mean Change from Baseline over Time}
\usage{
gen_ht_t14_310(adadas)
}
\arguments{
\item{adadas}{}
}
\value{
a huxtable
}
\description{
gen_ht_t14_310
Generate Table 14-3.10 ADAS Cog (11) - Mean and Mean Change from Baseline over Time
}
