% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTables.R
\name{gen_ht_t14_312}
\alias{gen_ht_t14_312}
\title{gen_ht_t14_312
Generate Table 14-3.12 Mean NPI-X Total Score from Week 4 through Week 24 - Windowed}
\usage{
gen_ht_t14_312(adnpix)
}
\arguments{
\item{adnpix}{}
}
\value{
a huxtable
}
\description{
gen_ht_t14_312
Generate Table 14-3.12 Mean NPI-X Total Score from Week 4 through Week 24 - Windowed
}
