% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTables.R
\name{gen_ht_t14_313}
\alias{gen_ht_t14_313}
\title{gen_ht_t14_313
Generate Table 14-3.13 CIBIC+ - Categorical Analysis - LOCF}
\usage{
gen_ht_t14_313(adadcibc)
}
\arguments{
\item{adadcibc}{}
}
\value{
a huxtable
}
\description{
gen_ht_t14_313
Generate Table 14-3.13 CIBIC+ - Categorical Analysis - LOCF
}
