% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTables.R
\name{gen_ht_t14_401}
\alias{gen_ht_t14_401}
\title{gen_ht_t14_401
Generate Table 14-4.01 Summary of Planned Exposure to Study Drug, as of End of Study}
\usage{
gen_ht_t14_401(adsl)
}
\arguments{
\item{adsl}{}
}
\value{
a huxtable
}
\description{
gen_ht_t14_401
Generate Table 14-4.01 Summary of Planned Exposure to Study Drug, as of End of Study
}
