% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTables.R
\name{gen_ht_t14_501}
\alias{gen_ht_t14_501}
\title{gen_ht_t14_501
Generate Table 14-5.01 Incidence of Treatment Emergent Adverse Events by Treatment Group}
\usage{
gen_ht_t14_501(adae, adsl)
}
\arguments{
\item{adsl}{}
}
\value{
a huxtable
}
\description{
gen_ht_t14_501
Generate Table 14-5.01 Incidence of Treatment Emergent Adverse Events by Treatment Group
}
