% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTables.R
\name{gen_ht_t14_502}
\alias{gen_ht_t14_502}
\title{gen_ht_t14_502
Generate Table 14-5.02 Incidence of Treatment Emergent Serious Adverse Events by Treatment Group}
\usage{
gen_ht_t14_502(adae, adsl)
}
\arguments{
\item{adae}{ADAE}

\item{adsl}{ADSL}
}
\value{
a huxtable
}
\description{
gen_ht_t14_502
Generate Table 14-5.02 Incidence of Treatment Emergent Serious Adverse Events by Treatment Group
}
\examples{
gen_ht_t14_502(adae = adae, adsl = adsl)

}
