% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTables.R
\name{gen_ht_t14_604}
\alias{gen_ht_t14_604}
\title{gen_ht_t14_604}
\usage{
gen_ht_t14_604(adlbc, adlbh, adsl)
}
\arguments{
\item{adlbc}{ADLBC}

\item{adlbh}{ADLBH}

\item{adsl}{ADSL}
}
\value{
a huxtable
}
\description{
Generate Table 14-6.04 Shifts of Laboratory Values During Treatment, Categorized Based on Threshold Ranges, by Visit
}
