% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTables.R
\name{gen_ht_t14_605}
\alias{gen_ht_t14_605}
\title{gen_ht_t14_605}
\usage{
gen_ht_t14_605(adlbc, adlbh, adsl)
}
\arguments{
\item{adsl}{}
}
\value{
a huxtable
}
\description{
Generate Table 14-6.05 Shifts of Laboratory Values During Treatment, Categorized Based on Threshold Ranges
}
