% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTables.R
\name{gen_ht_t14_701}
\alias{gen_ht_t14_701}
\title{gen_ht_t14_701
Generate Table 14-7.01 Summary of Vital Signs at Baseline and End of Treatment}
\usage{
gen_ht_t14_701(adsl, advs)
}
\arguments{
\item{adsl}{ADSL}

\item{advs}{ADVS}
}
\value{
a huxtable
}
\description{
gen_ht_t14_701
Generate Table 14-7.01 Summary of Vital Signs at Baseline and End of Treatment
}
