% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTables.R
\name{gen_ht_t14_704}
\alias{gen_ht_t14_704}
\title{gen_ht_t14_704
Generate Table 14-7.04 Summary of Concomitant Medications (Number of Subjects)}
\usage{
gen_ht_t14_704(sdtm_cm, adsl)
}
\arguments{
\item{sdtm_cm}{SDTM CM}

\item{adsl}{ADSL}
}
\value{
a haxtable
}
\description{
gen_ht_t14_704
Generate Table 14-7.04 Summary of Concomitant Medications (Number of Subjects)
}
