% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{invert.list}
\alias{invert.list}
\title{invert.list
Invert a list's items and names (assuming it's key value)}
\usage{
invert.list(NL)
}
\arguments{
\item{NL}{}
}
\description{
invert.list
Invert a list's items and names (assuming it's key value)
}
