% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_react.R
\name{lb_react}
\alias{lb_react}
\title{Create safety line plot}
\usage{
lb_react(data, paramVar, visVar, adyVar, avalVar, lowVar, highVar)
}
\arguments{
\item{data}{long data frame such as LB or VS}

\item{paramVar}{term column, PARAM}

\item{visVar}{analysis visit column, AVISIT}

\item{adyVar}{analysis day column, ADY}

\item{avalVar}{analysis value columnm, AVAL}

\item{lowVar}{lower limit column, A1LO}

\item{highVar}{upper limit column, A1HI}
}
\value{
an lineplot created with ggplot
}
\description{
Create safety line plot
}
