% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileApp.R
\name{profileApp}
\alias{profileApp}
\title{Profile App with AE and Labs data presentations}
\usage{
profileApp(
  data = list(dm = adam_adsl, aes = adam_adae, labs = adam_adlbc),
  settings = NULL,
  ptid = NULL,
  launchBrowser = FALSE,
  runNow = TRUE
)
}
\arguments{
\item{data}{ADAE, ADSL and ADLBC data}

\item{settings}{safetyGraphics settings}

\item{ptid}{participant ID to select when app is initialized}
}
\description{
Profile App with AE and Labs data presentations
}
