% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{sum_subgrp}
\alias{sum_subgrp}
\title{sum_subgrp}
\usage{
sum_subgrp(
  .data,
  subgroup_var,
  order_var = NULL,
  include.n = TRUE,
  pad.row = TRUE,
  header_n = header_n
)
}
\arguments{
\item{header_n}{}
}
\description{
sum_subgrp
}
