% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{summary_data}
\alias{summary_data}
\title{summary_data
Create the summary data portion for 14-3.01}
\usage{
summary_data(data, var, week, stub_header)
}
\arguments{
\item{stub_header}{}
}
\description{
summary_data
Create the summary data portion for 14-3.01
}
