% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTables.R
\name{gen_ht_t14_601}
\alias{gen_ht_t14_601}
\title{gen_ht_t14_601
Generate Table 14-6.01 Summary Statistics for Continuous Laboratory Values}
\usage{
gen_ht_t14_601(adlbc, adlbh)
}
\arguments{
\item{adlbc}{ADLBC}

\item{adlbh}{ADLBH}
}
\value{
a huxtable
}
\description{
gen_ht_t14_601
Generate Table 14-6.01 Summary Statistics for Continuous Laboratory Values
}
